% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_high_contrast.R
\name{scale_fill_high_contrast}
\alias{scale_fill_high_contrast}
\title{Define color fill for a bar chart.}
\usage{
scale_fill_high_contrast(foreground_color = "white")
}
\arguments{
\item{foreground_color}{fill color for bar (\code{"white"})}
}
\value{
Vector
}
\description{
Description goes here.
}
\examples{
scale_fill_high_contrast(foreground_color = "white")

}
