% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_linetype_high_contrast.R
\name{scale_linetype_high_contrast}
\alias{scale_linetype_high_contrast}
\title{Define linetypes for a line chart.}
\usage{
scale_linetype_high_contrast()
}
\value{
Vector
}
\description{
Description goes here.
}
\examples{
scale_linetype_high_contrast()

}
