% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_high_contrast.R
\name{theme_high_contrast}
\alias{theme_high_contrast}
\title{A high contrast theme.}
\usage{
theme_high_contrast(
  base_size = 11,
  base_family = "",
  foreground_color = "white",
  background_color = "black"
)
}
\arguments{
\item{base_size}{font size to use for the text in the plot (\code{11})}

\item{base_family}{font to use for the text in the plot (\code{"InputMono"})}

\item{foreground_color}{color for text and lines (\code{"white"})}

\item{background_color}{color for the background (\code{"black"})}
}
\value{
None
}
\description{
Description goes here.
}
\examples{
theme_high_contrast()

}
